% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colSBM_cluster_members}
\alias{colSBM_cluster_members}
\title{Extract from colSBM output, networks' assignment.}
\usage{
colSBM_cluster_members(cl)
}
\arguments{
\item{cl}{output of colSBM::clusterize_networks}
}
\value{
networks assignment
}
\description{
Extract from colSBM output, networks' assignment.
}
