% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colSBM_dendro_f}
\alias{colSBM_dendro_f}
\title{Extract from colSBM output, networks' assignment.}
\usage{
colSBM_dendro_f(tree, cl)
}
\arguments{
\item{tree}{empty list}

\item{cl}{output of colSBM::clusterize_networks}
}
\value{
dendrogram
}
\description{
Extract from colSBM output, networks' assignment.
}
\examples{
\dontrun{
tree = list()
tree = colSBM_dendro_f(tree,cl_col)
class(tree) = "dendrogram"
plot(rev(tree))
}
}
