% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colSBM_nodes_members}
\alias{colSBM_nodes_members}
\title{Extract from colSBM output, nodes' assignment.}
\usage{
colSBM_nodes_members(cl)
}
\arguments{
\item{cl}{output of colSBM::clusterize_networks}
}
\value{
nodes assignment
}
\description{
Extract from colSBM output, nodes' assignment.
}
