% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_perm_gamma}
\alias{dist_perm_gamma}
\title{Checking the distance between gamma and gamma with permutations
if the distance is low it is a sign of multiple class with the same behaviour,
there is a lot of limitation of this measure, especially low distance doesn't always mean harder to infer.}
\usage{
dist_perm_gamma(gamma_m)
}
\arguments{
\item{gamma_m}{Probability matrix for an edge between two nodes knowing their class.}
}
\value{
minimun of distances
}
\description{
Checking the distance between gamma and gamma with permutations
if the distance is low it is a sign of multiple class with the same behaviour,
there is a lot of limitation of this measure, especially low distance doesn't always mean harder to infer.
}
