% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exicl.R
\name{exICL_f}
\alias{exICL_f}
\title{Implementation of the exICL algorithm for SBM}
\usage{
exICL_f(N, Z_init, adj_m, return_exicl = FALSE)
}
\arguments{
\item{N}{Number of nodes}

\item{Z_init}{A binary matrix (K,N) defining class assignment.}

\item{adj_m}{Adjacency matrix.}

\item{return_exicl}{if true returns the exact ICL value for the inference.}
}
\value{
A binary matrix (K,N) defining class assignment.
}
\description{
Implementation of the exICL algorithm for SBM
}
