% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exicl.R
\name{exICL_with_rand_init}
\alias{exICL_with_rand_init}
\title{Generate random nodes' class assignment to initialize exICL algo}
\usage{
exICL_with_rand_init(N, Ncl_init, adj_m)
}
\arguments{
\item{N}{Number of nodes.}

\item{Ncl_init}{Number of nodes in each class.}

\item{adj_m}{Adjacency matrix.}
}
\value{
A binary matrix (K,N) defining class assignment.
}
\description{
Generate random nodes' class assignment to initialize exICL algo
}
