% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exicl.R
\name{exICL_with_sbm_init}
\alias{exICL_with_sbm_init}
\title{Generate nodes' class assignment from sbm package to initialize exICL algo}
\usage{
exICL_with_sbm_init(N, adj_m)
}
\arguments{
\item{N}{Number of nodes.}

\item{adj_m}{Adjacency matrix.}
}
\value{
A binary matrix (K,N) defining class assignment.
}
\description{
Generate nodes' class assignment from sbm package to initialize exICL algo
}
