% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_mix_sbm}
\alias{gen_mix_sbm}
\title{Compare package graphclust and colSBM to cluster a collection of SBM.}
\usage{
gen_mix_sbm(N, K, M, C, R, thetaMix = NULL)
}
\arguments{
\item{N}{Number of nodes in a graph.}

\item{K}{Maximal number of communities. (useless if thetaMix is given)}

\item{M}{Number of graphs.}

\item{C}{Number of clusters. There is, at least, one network by cluster. (useless if thetaMix is given)}

\item{R}{Number of independent tests.}

\item{thetaMix}{Optional. list of clusters properties. list(pi,gamma,prop)}
}
\value{
a dataframe with multiple properties on columns and each row is a test.
}
\description{
Compare package graphclust and colSBM to cluster a collection of SBM.
}
