% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_sbm}
\alias{gen_sbm}
\title{Compare different SBM inference techniques.}
\usage{
gen_sbm(N, R)
}
\arguments{
\item{N}{Number of nodes in a graph.}

\item{R}{Number of independent tests.}
}
\value{
a dataframe with multiple properties on columns and each row is a test.
}
\description{
Compare different SBM inference techniques.
}
