% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen_vfold}
\alias{gen_vfold}
\title{Generate graphs and apply on each the vfold algorithm.}
\usage{
gen_vfold(N, R)
}
\arguments{
\item{N}{Number of nodes in a graph.}

\item{R}{Number of graphs generated.}
}
\value{
a dataframe with multiple properties on columns and each row is a test.
}
\description{
Generate graphs and apply on each the vfold algorithm.
}
