% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsbm_sym_f}
\alias{rsbm_sym_f}
\title{Generate a non directed SBM}
\usage{
rsbm_sym_f(N, pi_v, gamma_m)
}
\arguments{
\item{N}{Number of nodes.}

\item{pi_v}{Probability vector to belong to a given class.}

\item{gamma_m}{Symmetric probability matrix for an edge between two nodes knowing their class.}
}
\value{
A binary matrix (K,N) defining class assignment and an adjacency matrix.
}
\description{
Generate a non directed SBM
}
