% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_to_num}
\alias{vec_to_num}
\title{Transform a vectorial representation of a class to a numbered representation}
\usage{
vec_to_num(vec)
}
\arguments{
\item{vec}{A binary matrix (K,N) defining class assignment.}
}
\value{
A numeric vector of size N.
}
\description{
Transform a vectorial representation of a class to a numbered representation
}
\examples{
\dontrun{
vec = matrix(c(1,0,0,1,1,0),2,3)
vec_to_num(vec)
c(1,2,1)
}
}
