% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vem.R
\name{vem_f}
\alias{vem_f}
\title{Implementation of the Variationnal Expectation-Maximization algorithm for SBM}
\usage{
vem_f(pi_v, gamma_m, adj_m)
}
\arguments{
\item{pi_v}{Probability vector to belong to a given class.}

\item{gamma_m}{Probability matrix for an edge between two nodes knowing their class.}

\item{adj_m}{Adjacency matrix.}
}
\value{
list of SBM parameters infered and nodes' class assignment
}
\description{
Implementation of the Variationnal Expectation-Maximization algorithm for SBM
}
