% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfold.R
\name{vfold}
\alias{vfold}
\title{Implementation of the vertices-fold algorithm for SBM}
\usage{
vfold(adj, v = 5, method = method_col)
}
\arguments{
\item{adj}{Adjacency matrix.}

\item{v}{Number of folds}

\item{method}{method to use to infer sbm method_col/method_gc corresponding to VEM/exICL.}
}
\value{
Average vfold error.
}
\description{
Implementation of the vertices-fold algorithm for SBM
}
