% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Z_to_theta}
\alias{Z_to_theta}
\title{Knowing nodes' class assignment and adjacency matrix, derive model parameters.}
\usage{
Z_to_theta(Z_m, adj_m)
}
\arguments{
\item{Z_m}{A binary matrix (K,N) defining class assignment.}

\item{adj_m}{Adjacency matrix.}
}
\value{
list of SBM parameters
}
\description{
Knowing nodes' class assignment and adjacency matrix, derive model parameters.
}
